requirements.txt
-e .
dash-mantine-components>=2.4.0
python-dotenv>=1.0.0
requests>=2.31.0
PyNaCl>=1.5.0
google-generativeai>=0.8.0
